requirements.txt
pathlib==1.0.1
py==1.4.34
gspread==3.0.0
oauth2client==4.1.2
pytest-flake8==1.0.0
pytest-cov==2.5.1
ansicolors==1.1.8
numpy==1.14.2
pandas==0.22.0
